+++ test/trace/debugger_test.rb
@@ -50,7 +50,7 @@ class DebuggerTest < Minitest::Spec
     )
 
     assert_equal debugger_nodes[0].task, activity
-    assert_equal debugger_nodes[0].instance_variable_get(:@activity), Trailblazer::Activity::TaskWrap.container_activity_for(activity)
+    assert_equal debugger_nodes[0].activity, Trailblazer::Activity::TaskWrap.container_activity_for(activity)
 
     assert_equal debugger_nodes[0].task, activity
     assert_equal debugger_nodes[0].compile_id, nil
@@ -63,7 +63,7 @@ class DebuggerTest < Minitest::Spec
     assert_equal debugger_nodes[0].captured_input, stack.to_a[0]
     assert_equal debugger_nodes[0].captured_output, stack.to_a[-1]
 
-    assert_equal debugger_nodes[1].instance_variable_get(:@activity).class, Trailblazer::Activity # The [activity] field is an Activity.
+    assert_equal debugger_nodes[1].activity.class, Trailblazer::Activity # The [activity] field is an Activity.
     assert_equal debugger_nodes[1].task.inspect, %{#<Trailblazer::Activity::Start semantic=:default>}
     assert_equal debugger_nodes[1].compile_id, %{Start.default}
     assert_equal debugger_nodes[1].compile_path, ["Start.default"]
@@ -77,8 +77,8 @@ class DebuggerTest < Minitest::Spec
 
     assert_equal debugger_nodes[3].task, sub_activity
     # the "parent activity" for {sub_activity} is not the Activity::Railway class but instance of Acivity.
-    assert_equal debugger_nodes[3].instance_variable_get(:@activity).class, Trailblazer::Activity
-    assert_equal debugger_nodes[3].instance_variable_get(:@activity), debugger_nodes[1].instance_variable_get(:@activity)
+    assert_equal debugger_nodes[3].activity.class, Trailblazer::Activity
+    assert_equal debugger_nodes[3].activity, debugger_nodes[1].activity
     assert_equal debugger_nodes[3].data, {exception_source: true}
     assert_equal debugger_nodes[3].runtime_path, ["B"]